a11y: spec testing/refactoring
